fixup relative addresses. still need to figure out the tiny thing that's causing this to panic.
if we are theoretically stuck at the instruction we delete, which is

ffffff80005369ef         mov        cl, byte [ds:rax+r15]

then we write a jmp to our function, which is five bytes total, theoretically we need to use this new address for every jmp/je executed afterwards in the function.

in other words we are 'frozen in time' after executing the jump, as once the system swallows those 5 bytes we now can do whatever we want.

something is missing. it's probably small.