offsets are wrong. we are not "frozen in time". i didn't realise we did jump from a spot, to another spot.
so we need to compute the offsets properly and with respect to the absolute location we want to hit.

@krackers found thsi great link:

https://stackoverflow.com/questions/61833681/get-a-label-address-out-of-the-function-scope-in-gcc-clang-c

i still have to figure out if i'm doing it right. probably not. i doubt we can just jam an absolute address and subtract it from a label for it to work.