just in case the neckbeard thinks its wrong (it's not), this change makes no difference.
i think using a trampoline for fdexec is too slow. will probably have to make a new kernel build entirely, which may be a good thing.

this change, in theory, did fix the panic. but there are other things, i suspect, that are still slowing down the kernel in comparison to 10.8.